+++ app/Components/Model/Model.php
@@ -9,7 +9,7 @@ abstract class Model
     protected $primaryKey;
     protected $columns;
 
-    protected function instance()
+    protected function database()
     {
         $katrina = new ORM($this->table, $this->primaryKey, $this->columns);
         return $katrina;