+++ config/config.php
@@ -19,3 +19,8 @@ define('INDEX_LOGIN', 'solital_index_login');
 
 define('SECRET_IV', pack('a16', 'first_secret'));
 define('SECRET', pack('a16', 'second_secret'));
+
+/**
+ * MONOLOG DIRECTORY
+ */
+define('MONOLOG_DIR', dirname(__DIR__).'/app/LogFiles/');